Use LKG data if OptProf-data-collection tests fail (#12916)
This should improve the stability of our builds by ensuring that all
optprof data drops have data in them--as opposed to now, when a failed
test can cause an empty drop, causing subsequent builds to fail.

All links below are internal:

Example optprof collection job that failed but fell back to LKG data:
https://dev.azure.com/devdiv/DevDiv/DevDiv%20Team/_build/results?buildId=12897383&view=logs&j=195fae89-8926-5512-4513-4f434154ec35&t=44ab5a05-4136-570a-3ce3-f35d63e9ae60&l=582

Using that data to complete the build:
https://dev.azure.com/devdiv/DevDiv/DevDiv%20Team/_build/results?buildId=12932643&view=logs&j=bb592630-4b9d-53ad-3960-d954a70a95cf&t=94418e61-6648-5751-f7d4-a14f4e5e2bb7&l=36

Docs on the LKG system:
https://devdiv.visualstudio.com/DevDiv/_wiki/wikis/DevDiv.wiki/29053/Enabling-LKG-support

Because we don't use the MicroBuild plugin to acquire the data, I had to
change a couple of things from the defaults:

* Add the LKG data to a new copy of the `Metadata.json` file after
Arcade finds the drop in the target `UpdateOptProfMetadataJson`. Had to
be a new copy of the file because `MicroBuildSigningPlugin` uploads that
directly. Fortunately that was fine because I was also able to
* Change the download + data-extraction script to handle the different
artifact.